requirements.txt
anthropic
fastapi
uvicorn[standard]
rank-bm25
pydantic
python-dotenv
# Phase 2 (uncomment when needed):
# sentence-transformers
# lancedb
